% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{adiv_binary_normalized}
\alias{adiv_binary_normalized}
\title{documentation}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features)
or a numeric vector representing a single sample. Counts are
automatically converted to presence/absence (binary) values prior to
calculation.}
}
\description{
documentation
}
\keyword{internal}
