% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{adiv_percent_normalized}
\alias{adiv_percent_normalized}
\title{documentation}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features)
or a numeric vector representing a single sample. Raw counts are
automatically converted to relative abundances (proportions).}
}
\description{
documentation
}
\keyword{internal}
