% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{bdiv_assert_integer}
\alias{bdiv_assert_integer}
\title{documentation}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Values must be integers (non-integer counts will return an error).}
}
\description{
documentation
}
\keyword{internal}
