% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{bdiv_binary_normalized}
\alias{bdiv_binary_normalized}
\title{documentation}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Counts are automatically converted to presence/absence (binary)
values prior to calculation.}
}
\description{
documentation
}
\keyword{internal}
