% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{bdiv_percent_normalized}
\alias{bdiv_percent_normalized}
\title{documentation}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Counts are automatically converted to relative abundances (proportions
summing to 1) prior to calculation.}
}
\description{
documentation
}
\keyword{internal}
