% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{beta_div}
\alias{beta_div}
\title{Beta diversity wrapper}
\usage{
beta_div(counts, metric, ...)
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{metric}{The name of a beta diversity metric. Current options are
\code{c('bray_curtis', 'sorenson', 'canberra', 'euclidean', 'gower', 'jaccard', 'kulczynski', 'manhattan', 'unweighted_unifrac', 'weighted_unifrac', 'normalized_unifrac', 'generalized_unifrac', 'variance_adjusted_unifrac')}.
Supports flexible name matching - see details section below. Options are
also available via \code{names(metrics$beta)}.}

\item{...}{Additional options to pass through to the called function. I.e.
\code{tree}, \code{weighted}, \code{pairs}, \code{alpha}, or \code{cpus}.}
}
\value{
A numeric vector.
}
\description{
Beta diversity wrapper
}
\details{
\strong{Flexible name matching}

Case insensitive and partial matching. Any runs of non-alpha characters are
converted to underscores. E.g. \verb{metric = 'Bray Curtis} selects Bray-Curtis.

UniFrac names can be shortened to the first letter plus "unifrac". E.g.
\code{uunifrac}, \code{w_unifrac}, or \verb{V UniFrac}. These also support partial matching,
but the primary options take precedence. For instance \code{metric = 'g'}
selects Gower, whereas \code{metric = 'gu'} selects Generalized UniFrac.

Finished code should always use the full primary option name to avoid
ambiguity with future additions to the metrics list.
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Bray-Curtis distances
    beta_div(ex_counts, 'bray_curtis')
    
    # Kulczynski distances
    beta_div(ex_counts, 'k')
    
    # Generalized UniFrac distances
    beta_div(ex_counts, 'GUniFrac', tree = ex_tree)
    
    
}
\seealso{
Other alpha_diversity: 
\code{\link{alpha_div}()},
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{inv_simpson}()},
\code{\link{observed}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
