% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{brillouin}
\alias{brillouin}
\title{Brillouin Index}
\usage{
brillouin(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A diversity index derived from information theory, appropriate for fully censused communities.
}
\details{
The Brillouin index is defined as:
\deqn{\frac{\ln{[(\sum_{i = 1}^{n} X_i)!]} - \sum_{i = 1}^{n} \ln{(X_i!)}}{\sum_{i = 1}^{n} X_i}}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{X_i} : Integer count of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# note: lgamma(x + 1) == log(x!)
(lgamma(sum(x) + 1) - sum(lgamma(x + 1))) / sum(x)
}\if{html}{\out{</div>}}
}
\examples{
    brillouin(ex_counts)
}
\references{
Brillouin, L. (1956). Science and information theory. Academic Press.
}
\seealso{
alpha_div

Other Diversity metrics: 
\code{\link{fisher}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{Diversity metrics}
