% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{chao1}
\alias{chao1}
\title{Chao1 Richness Estimator}
\usage{
chao1(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A non-parametric estimator of the lower bound of species richness.
}
\details{
The Chao1 estimator uses the ratio of singletons to doubletons to estimate the number of missing species:
\deqn{n + \frac{(F_1)^2}{2 F_2}}

Where:
\itemize{
\item \eqn{n} : The number of observed features.
\item \eqn{F_1} : Number of features observed once (singletons).
\item \eqn{F_2} : Number of features observed twice (doubletons).
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum(x>0) + (sum(x == 1) ** 2) / (2 * sum(x == 2))
}\if{html}{\out{</div>}}
}
\examples{
    chao1(ex_counts)
}
\references{
Chao, A. (1984). Nonparametric estimation of the number of classes in a population. \emph{Scandinavian Journal of Statistics}, 11, 265-270.
}
\seealso{
alpha_div

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{observed}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
