% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_counts}
\alias{ex_counts}
\title{Example counts matrix}
\format{
A matrix of 4 samples (column) x 5 OTUs (rows).
}
\usage{
ex_counts
}
\description{
Example counts matrix
}
\keyword{datasets}
