% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_tree}
\alias{ex_tree}
\title{Example phylogenetic tree}
\format{
A \code{phylo} object.
}
\usage{
ex_tree
}
\description{
Companion tree for \code{ex_counts}.
}
\details{
\code{ex_tree} encodes this tree structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{      +----------44---------- Haemophilus
  +-2-|
  |   +----------------68---------------- Bacteroides  
  |                      
  |             +---18---- Streptococcus
  |      +--12--|       
  |      |      +--11-- Staphylococcus
  +--11--|              
         |      +-----24----- Corynebacterium
         +--12--|
                +--13-- Propionibacterium
}\if{html}{\out{</div>}}
}
\keyword{datasets}
