% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{fisher}
\alias{fisher}
\title{Fisher's Alpha}
\usage{
fisher(counts, digits = 3L, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{digits}{Precision of the returned values, in number of decimal
places. E.g. the default \code{digits=3} could return \code{6.392}.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A parametric diversity index assuming species abundances follow a log-series distribution.
}
\details{
Fisher's Alpha (\eqn{\alpha}) is the parameter in the equation:
\deqn{\frac{n}{\alpha} = \ln{\left(1 + \frac{X_T}{\alpha}\right)}}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{X_T} : Total of all counts (sequencing depth).
}

The value of \eqn{\alpha} is solved for iteratively.

\strong{Parameter: digits}

The precision (number of decimal places) to use when solving the equation.
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    fisher(ex_counts)
}
\references{
Fisher, R. A., Corbet, A. S., & Williams, C. B. (1943). The relation between the number of species and the number of individuals in a random sample of an animal population. \emph{Journal of Animal Ecology}, 12, 42-58. \doi{10.2307/1411}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{Diversity metrics}
