% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{hamming}
\alias{hamming}
\title{Hamming distance}
\usage{
hamming(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Measures the minimum number of substitutions required to change one string into the other.
}
\details{
The Hamming distance is defined as:
\deqn{(A + B) - 2J}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum(xor(x, y))
}\if{html}{\out{</div>}}
}
\examples{
    hamming(ex_counts)
}
\references{
Hamming, R. W. (1950). Error detecting and error correcting codes. \emph{Bell System Technical Journal}, 29(2), 147-160. \doi{10.1002/j.1538-7305.1950.tb00463.x}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Presence/Absence metrics: 
\code{\link{jaccard}()},
\code{\link{ochiai}()},
\code{\link{sorensen}()}
}
\concept{Presence/Absence metrics}
