% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{jaccard}
\alias{jaccard}
\title{Jaccard distance}
\usage{
jaccard(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Measures dissimilarity between sample sets.
}
\details{
The Jaccard distance is defined as:
\deqn{1 - \frac{J}{(A + B - J)}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{1 - sum(x & y) / sum(x | y)
}\if{html}{\out{</div>}}
}
\examples{
    jaccard(ex_counts)
}
\references{
Jaccard, P. (1912). The distribution of the flora in the alpine zone. \emph{New Phytologist}, 11(2), 37-50. \doi{10.1111/j.1469-8137.1912.tb05611.x}
}
\seealso{
beta_div

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{ochiai}()},
\code{\link{sorensen}()}
}
\concept{Presence/Absence metrics}
