% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{jsd}
\alias{jsd}
\title{Jensen-Shannon divergence (JSD)}
\usage{
jsd(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A symmetrized and smoothed version of the Kullback-Leibler divergence.
}
\details{
The Jensen-Shannon divergence (JSD) is defined as:
\deqn{\frac{1}{2}\left[\sum_{i=1}^{n}P_i\ln\left(\frac{2P_i}{P_i + Q_i}\right) + \sum_{i=1}^{n}Q_i\ln\left(\frac{2Q_i}{P_i + Q_i}\right)\right]}

Where:
\itemize{
\item \eqn{P_i}, \eqn{Q_i} : Proportional abundances of the \eqn{i}-th feature.
\item \eqn{n} : The number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]; p <- x / sum(x)
y <- ex_counts[2,]; q <- y / sum(y)
sum(p * log(2 * p / (p+q)), q * log(2 * q / (p+q))) / 2
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    jsd(ex_counts)
}
\references{
Lin, J. (1991). Divergence measures based on the Shannon entropy. \emph{IEEE Transactions on Information Theory}, 37(1), 145-151. \doi{10.1109/18.61115}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
