% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{margalef}
\alias{margalef}
\title{Margalef's Richness Index}
\usage{
margalef(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A richness metric that normalizes the number of species by the log of the total sample size.
}
\details{
Margalef's index is defined as:
\deqn{\frac{n - 1}{\ln{X_T}}}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{X_T} : Total of all counts (sequencing depth).
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
(sum(x > 0) - 1) / log(sum(x))
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    margalef(ex_counts)
}
\references{
Margalef, R. (1958). Information theory in ecology. \emph{General Systems}, 3, 36-71.

Gamito, S. (2010). Caution is needed when applying Margalef diversity index. \emph{Ecological Indicators}, 10(2), 550-551. \doi{10.1016/j.ecolind.2009.07.006}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{chao1}()},
\code{\link{menhinick}()},
\code{\link{observed}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
