% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{mcintosh}
\alias{mcintosh}
\title{McIntosh Index}
\usage{
mcintosh(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A dominance index based on the Euclidean distance from the origin.
}
\details{
The McIntosh index is defined as:
\deqn{\frac{X_T - \sqrt{\sum_{i = 1}^{n} (X_i)^2}}{X_T - \sqrt{X_T}}}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{X_i} : Integer count of the \eqn{i}-th feature.
\item \eqn{X_T} : Total of all counts.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
(sum(x) - sqrt(sum(x^2))) / (sum(x) - sqrt(sum(x)))
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    mcintosh(ex_counts)
}
\references{
McIntosh, R. P. (1967). An index of diversity and the relation of certain concepts to diversity. \emph{Ecology}, 48(3), 392-404. \doi{10.2307/1932674}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Dominance metrics: 
\code{\link{berger}()}
}
\concept{Dominance metrics}
