% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{morisita}
\alias{morisita}
\title{Morisita dissimilarity}
\usage{
morisita(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A measure of overlap between samples that is independent of sample size. Requires integer counts.
}
\details{
The Morisita dissimilarity is defined as:
\deqn{1 - \frac{2\sum_{i=1}^{n}X_{i}Y_{i}}{\left(\frac{\sum_{i=1}^{n}X_i(X_i - 1)}{X_T(X_T - 1)} + \frac{\sum_{i=1}^{n}Y_i(Y_i - 1)}{Y_T(Y_T - 1)}\right)X_{T}Y_{T}}}

Where:
\itemize{
\item \eqn{X_i}, \eqn{Y_i} : Absolute counts of the \eqn{i}-th feature.
\item \eqn{X_T}, \eqn{Y_T} : Total counts in each sample. \eqn{X_T = \sum_{i=1}^{n} X_i}.
\item \eqn{n} : The number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{simp_x <- sum(x * (x - 1)) / (sum(x) * (sum(x) - 1))
simp_y <- sum(y * (y - 1)) / (sum(y) * (sum(y) - 1))
1 - ((2 * sum(x * y)) / ((simp_x + simp_y) * sum(x) * sum(y)))
}\if{html}{\out{</div>}}
}
\examples{
    morisita(ex_counts)
}
\references{
Morisita, M. (1959). Measuring of interspecific association and similarity between communities. \emph{Memoirs of the Faculty of Science, Kyushu University, Series E (Biology)}, 3, 65-80.
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
