% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{n_cpus}
\alias{n_cpus}
\title{Number of CPU Cores}
\usage{
n_cpus()
}
\value{
A scalar integer, guaranteed to be at least \code{1}.
}
\description{
A thin wrapper around \code{parallely::availableCores()}. If the \code{parallely}
package is not installed, then it falls back to
\code{parallel::detectCores(all.tests = TRUE, logical = TRUE)}. Returns \code{1} if
\code{pthread} support is unavailable or when the number of cpus cannot be
determined.
}
\examples{
    n_cpus()

}
