% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{normalized_unifrac}
\alias{normalized_unifrac}
\title{Normalized Weighted UniFrac}
\usage{
normalized_unifrac(
  counts,
  tree = NULL,
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
must be present in \code{tree}. Can be omitted if a tree is embedded with
the \code{counts} object or as \code{attr(counts, 'tree')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Weighted UniFrac normalized by the tree length to allow comparison between trees.
}
\details{
The Normalized Weighted UniFrac distance is defined as:
\deqn{\frac{\sum_{i=1}^{n} L_i|P_i - Q_i|}{\sum_{i=1}^{n} L_i(P_i + Q_i)}}

Where:
\itemize{
\item \eqn{n} : The number of branches in the tree.
\item \eqn{L_i} : The length of the \eqn{i}-th branch.
\item \eqn{P_i}, \eqn{Q_i} : The proportion of the community descending from branch \eqn{i} in sample P and Q.
}
}
\examples{
    normalized_unifrac(ex_counts, tree = ex_tree)
}
\references{
Lozupone, C. A., Hamady, M., Kelley, S. T., & Knight, R. (2007). Quantitative and qualitative beta diversity measures lead to different insights into factors that structure microbial communities. \emph{Applied and Environmental Microbiology}, 73(5), 1576-1585. \doi{10.1128/AEM.01996-06}
}
\seealso{
beta_div

Other Phylogenetic metrics: 
\code{\link{faith}()},
\code{\link{generalized_unifrac}()},
\code{\link{unweighted_unifrac}()},
\code{\link{variance_adjusted_unifrac}()},
\code{\link{weighted_unifrac}()}
}
\concept{Phylogenetic metrics}
