% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{observed}
\alias{observed}
\title{Observed Features}
\usage{
observed(counts, cpus = n_cpus())
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\value{
A numeric vector.
}
\description{
Observed Features alpha diversity metric.
}
\details{
This is the most straightforward and intuitive measure of diversity. It is a
simple count of the number of unique microbial taxa (such as Amplicon
Sequence Variants, or ASVs) detected in a sample. A higher value indicates
greater richness. While easy to understand, this metric is highly sensitive
to the number of sequences per sample (sequencing depth). A sample with more
sequences is more likely to detect rare taxa by chance, leading to an
inflated richness value. Therefore, it is not appropriate to directly compare
the Observed Features of samples with different sequencing depths without
first normalizing the data, typically through a process called rarefaction
(subsampling all samples to an equal depth).
}
\section{Calculation}{


Pre-transformation: drop all OTUs with zero abundance.

In the formulas below, \eqn{x} is a single column (sample) from \code{counts}.

\deqn{p_{i} = \displaystyle \frac{x_i}{\sum x}}
\deqn{D = \displaystyle \sum_{i = 1}^{n} 1}

\if{html}{\out{<div class="sourceCode">}}\preformatted{  x <- c(4, 0, 3, 2, 6)[-2]
  length(x)
  #>  4
}\if{html}{\out{</div>}}
}

\examples{
    # Example counts matrix
    ex_counts
    
    # Observed features
    observed(ex_counts)
    
}
\seealso{
Other alpha_diversity: 
\code{\link{alpha_div}()},
\code{\link{beta_div}()},
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
