% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{robust_aitchison}
\alias{robust_aitchison}
\title{Robust Aitchison distance}
\usage{
robust_aitchison(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Calculates the pairwise Robust Aitchison distance for compositional data.
This method is specifically engineered for sparse datasets - such as microbiome
OTU/ASV tables - by calculating distances based only on observed positive
abundances, avoiding the need for pseudo-counts.
}
\details{
The Robust Aitchison distance is defined as:
\deqn{\sqrt{\sum_{i=1}^{n} (X^*_i - Y^*_i)^2}}

Where:
\itemize{
\item \eqn{X^*_i}, \eqn{Y^*_i} : The rclr-transformed counts for the \eqn{i}-th feature.
For a given sample \eqn{X}, \eqn{X^*_i = \ln(X_i) - X_L} if \eqn{X_i > 0}, and \eqn{0} otherwise.
\item \eqn{X_L}, \eqn{Y_L} : Mean log of strictly positive abundances.
\eqn{X_L = \frac{1}{|P_X|}\sum_{j \in P_X} \ln{X_j}}, where \eqn{P_X} is the set of indices where \eqn{X > 0}.
\item \eqn{|P_X|}, \eqn{|P_Y|} : The number of strictly positive features in the respective samples.
\item \eqn{n} : The total number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ifelse(x > 0, log(x) - mean(log(x[x > 0])), 0)
y <- ifelse(y > 0, log(y) - mean(log(y[y > 0])), 0)
sqrt(sum((x-y)^2)) # Euclidean distance
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    robust_aitchison(ex_counts)
}
\references{
Martino, C., Morton, J. T., Marotz, C. A., Thompson, L. R., Tripathi, A., Knight, R., and Zengler, K. (2019). A novel sparse compositional technique reveals microbial perturbations. \emph{mSystems}, 4(1), e00016-19. \doi{10.1128/mSystems.00016-19}
}
\seealso{
\code{beta_div()}, \code{aitchison()}

\code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
