% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{simpson}
\alias{simpson}
\title{Simpson alpha diversity metric}
\usage{
simpson(counts, cpus = n_cpus())
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Gauges the uniformity of species within a community. A Simpson index of \code{1}
indicates that one or a few high abundance OTUs dominate the community,
which is indicative of low diversity.
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Simpson diversity values
    simpson(ex_counts)
    
}
\seealso{
Other alpha_diversity: 
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()}
}
\concept{alpha_diversity}
