% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{squared_chisq}
\alias{squared_chisq}
\title{Squared Chi-Squared distance}
\usage{
squared_chisq(
  counts,
  norm = "percent",
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{norm}{Normalize the incoming counts. Options are:
* \code{'percent'}: Relative abundance (sample abundances sum to 1).
* \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
* \code{'clr'}: Centered log ratio.
* \code{'none'}: No transformation.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   Default: `'percent'`, which is the expected input for these formulas.
}\if{html}{\out{</div>}}}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The squared version of the Chi-Squared distance.
}
\details{
The Squared \eqn{\chi^2} distance is defined as:
\deqn{\sum_{i=1}^{n}\frac{(P_i - Q_i)^2}{P_i + Q_i}}

Where:
\itemize{
\item \eqn{P_i}, \eqn{Q_i} : Proportional abundances of the \eqn{i}-th feature.
\item \eqn{n} : The number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum((x - y)^2 / (x + y))
}\if{html}{\out{</div>}}
}
\examples{
    squared_chisq(ex_counts)
}
\references{
Cha, S.-H. (2007). Comprehensive survey on distance/similarity measures between probability density functions. \emph{International Journal of Mathematical Models and Methods in Applied Sciences}, 1(4), 300–307.
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
