% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{squared_euclidean}
\alias{squared_euclidean}
\title{Squared Euclidean distance}
\usage{
squared_euclidean(
  counts,
  norm = "none",
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{norm}{Normalize the incoming counts. Options are:
\itemize{
\item \code{'none'}: No transformation.
\item \code{'percent'}: Relative abundance (sample abundances sum to 1).
\item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
\item \code{'clr'}: Centered log ratio.
}

Default: \code{'none'}.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The squared Euclidean distance between two vectors.
}
\details{
The Squared Euclidean distance is defined as:
\deqn{\sum_{i=1}^{n} (X_i - Y_i)^2}

Where:
\itemize{
\item \eqn{X_i}, \eqn{Y_i} : Absolute abundances of the \eqn{i}-th feature.
\item \eqn{n} : The number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum((x-y)^2)
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    squared_euclidean(ex_counts)
}
\references{
Legendre, P., & Legendre, L. (2012). Numerical ecology. Elsevier.
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
