% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{ace}
\alias{ace}
\title{Abundance-based Coverage Estimator (ACE)}
\usage{
ace(counts, cutoff = 10L, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{cutoff}{The maximum number of observations to consider "rare".
Default: \code{10}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A non-parametric estimator of species richness that separates features into abundant and rare groups.
}
\details{
The ACE metric separates features into "abundant" and "rare" groups based on a cutoff (usually 10 counts). It assumes that the presence of abundant species is certain, while the true number of rare species must be estimated.

\strong{Equations:}

\deqn{C_{ace} = 1 - \frac{F_1}{X_{rare}}}

\deqn{\gamma_{ace}^2 = \max\left[\frac{F_{rare} \sum_{i=1}^{r}i(i-1)F_i}{C_{ace}X_{rare}(X_{rare} - 1)} - 1, 0\right]}

\deqn{D_{ace} = F_{abund} + \frac{F_{rare}}{C_{ace}} + \frac{F_1}{C_{ace}}\gamma_{ace}^2}

Where:
\itemize{
\item \eqn{r} : Rare cutoff (default 10). Features with \eqn{\le r} counts are considered rare.
\item \eqn{F_i} : Number of features with exactly \eqn{i} counts.
\item \eqn{F_1} : Number of features where \eqn{X_i = 1} (singletons).
\item \eqn{F_{rare}} : Number of rare features where \eqn{X_i \le r}.
\item \eqn{F_{abund}} : Number of abundant features where \eqn{X_i > r}.
\item \eqn{X_{rare}} : Total counts belonging to rare features.
\item \eqn{C_{ace}} : The sample abundance coverage estimator.
\item \eqn{\gamma_{ace}^2} : The estimated coefficient of variation.
}

\strong{Parameter: cutoff}
The integer threshold distinguishing rare from abundant species. Standard practice is to use 10.
}
\examples{
    ace(ex_counts)
}
\references{
Chao, A., & Lee, S. M. (1992). Estimating the number of classes via sample coverage. \emph{Journal of the American Statistical Association}, 87(417), 210-217. \doi{10.1080/01621459.1992.10475194}
}
\seealso{
alpha_div

Other Richness metrics: 
\code{\link{chao1}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{observed}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
