% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{aitchison}
\alias{aitchison}
\title{Aitchison distance}
\usage{
aitchison(
  counts,
  pseudocount = NULL,
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{pseudocount}{The value to add to all counts in \code{counts} to prevent
taking \code{log(0)} for unobserved features. The default, \code{NULL}, selects
the smallest non-zero value in \code{counts}.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Calculates the Euclidean distance between centered log-ratio (CLR) transformed abundances.
}
\details{
The Aitchison distance is defined as:
\deqn{\sqrt{\sum_{i=1}^{n} [(\ln{X_i} - X_L) - (\ln{Y_i} - Y_L)]^2}}

Where:
\itemize{
\item \eqn{X_i}, \eqn{Y_i} : Absolute counts for the \eqn{i}-th feature.
\item \eqn{X_L}, \eqn{Y_L} : Mean log of abundances. \eqn{X_L = \frac{1}{n}\sum_{i=1}^{n} \ln{X_i}}.
\item \eqn{n} : The number of features.
}

\strong{Parameter: pseudocount}

Because the formula uses logarithms, zeros in the data must be handled.
The \code{pseudocount} argument adds a small value to all counts prior to calculation.

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- log((x + pseudocount) / exp(mean(log(x + pseudocount))))
y <- log((y + pseudocount) / exp(mean(log(y + pseudocount))))
sqrt(sum((x-y)^2)) # Euclidean distance
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    aitchison(ex_counts, pseudocount = 1)
}
\references{
Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall. \doi{10.1007/978-94-009-4109-3}

Aitchison, J. (1982). The statistical analysis of compositional data. \emph{Journal of the Royal Statistical Society: Series B (Methodological)}, 44(2), 139-160. \doi{10.1111/j.2517-6161.1982.tb01195.x}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{psym_chisq}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
