% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{alpha_div}
\alias{alpha_div}
\title{Alpha diversity wrapper}
\usage{
alpha_div(counts, metric, ...)
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{metric}{The name of an alpha diversity metric. Current options are
\code{c('observed', 'chao1', 'shannon', 'simpson', 'inv_simpson', 'faith')}.
Supports partial name matching. Options are also available via
\code{names(metrics$alpha)}.}

\item{...}{Additional options to pass through to the called function. I.e.
\code{cpus} or \code{tree}.}
}
\value{
A numeric vector.
}
\description{
Alpha diversity wrapper
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Shannon diversity values
    alpha_div(ex_counts, 'Shannon')
    
    # Chao1 diversity values
    alpha_div(ex_counts, 'c')
    
    # Faith PD values
    alpha_div(ex_counts, 'faith', tree = ex_tree)
    
    
}
\seealso{
Other alpha_diversity: 
\code{\link{beta_div}()},
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{inv_simpson}()},
\code{\link{observed}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
