% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{berger}
\alias{berger}
\title{Berger-Parker Index}
\usage{
berger(counts, norm = "percent", margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{norm}{Normalize the incoming counts. Options are:
\itemize{
\item \code{'percent'}: Relative abundance (sample abundances sum to 1).
\item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
\item \code{'clr'}: Centered log ratio.
\item \code{'none'}: No transformation.
}

Default: \code{'percent'}, which is the expected input for these formulas.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A measure of the numerical importance of the most abundant species.
}
\details{
The Berger-Parker index is defined as the proportional abundance of the most dominant feature:
\deqn{\max(P_i)}

Where:
\itemize{
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{max(x / sum(x))
}\if{html}{\out{</div>}}
}
\examples{
    berger(ex_counts)
}
\references{
Berger, W. H., & Parker, F. L. (1970). Diversity of planktonic foraminifera in deep-sea sediments. \emph{Science}, 168(3937), 1345-1347. \doi{10.1126/science.168.3937.1345}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Dominance metrics: 
\code{\link{mcintosh}()}
}
\concept{Dominance metrics}
