% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{berger}
\alias{berger}
\title{Berger-Parker Index}
\usage{
berger(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A measure of the numerical importance of the most abundant species.
}
\details{
The Berger-Parker index is defined as the proportional abundance of the most dominant feature:
\deqn{\max(P_i)}

Where:
\itemize{
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
p <- x / sum(x)
max(p)
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    berger(ex_counts)
}
\references{
Berger, W. H., & Parker, F. L. (1970). Diversity of planktonic foraminifera in deep-sea sediments. \emph{Science}, 168(3937), 1345-1347. \doi{10.1126/science.168.3937.1345}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Dominance metrics: 
\code{\link{mcintosh}()}
}
\concept{Dominance metrics}
