% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{beta_div}
\alias{beta_div}
\title{Beta diversity wrapper}
\usage{
beta_div(counts, metric, ...)
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{metric}{The name of a beta diversity metric. Current options are
\code{c('bray_curtis', 'sorenson', 'canberra', 'euclidean', 'gower', 'jaccard', 'kulczynski', 'manhattan', 'unweighted_unifrac', 'weighted_unifrac', 'weighted_normalized_unifrac', 'generalized_unifrac', 'variance_adjusted_unifrac')}. Supports partial name matching. Options are
also available via \code{names(metrics$beta)}.}

\item{...}{Additional options to pass through to the called function. I.e.
\code{tree}, \code{weighted}, \code{pairs}, \code{alpha}, or \code{cpus}.}
}
\value{
A numeric vector.
}
\description{
Beta diversity wrapper
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Shannon diversity values
    alpha_div(ex_counts, 'Shannon')
    
    # Chao1 diversity values
    alpha_div(ex_counts, 'c')
    
    # Faith PD values
    alpha_div(ex_counts, 'faith', tree = ex_tree)
    
    
}
\seealso{
Other alpha_diversity: 
\code{\link{alpha_div}()},
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{inv_simpson}()},
\code{\link{observed}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
