% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{beta_div}
\alias{beta_div}
\title{Beta Diversity Wrapper Function}
\usage{
beta_div(
  counts,
  metric,
  norm = "percent",
  power = 1.5,
  pseudocount = NULL,
  alpha = 0.5,
  tree = NULL,
  pairs = NULL,
  margin = 1L,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{metric}{The name of a beta diversity metric. One of \code{c('aitchison', 'bhattacharyya', 'bray', 'canberra', 'chebyshev', 'chord', 'clark', 'divergence', 'euclidean', 'generalized_unifrac', 'gower', 'hamming', 'hellinger', 'horn', 'jaccard', 'jensen', 'jsd', 'lorentzian', 'manhattan', 'matusita', 'minkowski', 'morisita', 'motyka', 'normalized_unifrac', 'ochiai', 'psym_chisq', 'soergel', 'sorensen', 'squared_chisq', 'squared_chord', 'squared_euclidean', 'topsoe', 'unweighted_unifrac', 'variance_adjusted_unifrac', 'wave_hedges', 'weighted_unifrac')}. Flexible
matching is supported (see below). Programmatic access via
\code{list_metrics('beta')}.}

\item{norm}{Normalize the incoming counts. Options are:
\describe{
\item{\code{norm = "percent"} - }{ Relative abundance (sample abundances sum to 1). }
\item{\code{norm = "binary"}  - }{ Unweighted presence/absence (each count is either 0 or 1). }
\item{\code{norm = "clr"}     - }{ Centered log ratio. }
\item{\code{norm = "none"}    - }{ No transformation. }
}
Default: \code{'percent'}, which is the expected input for these formulas.}

\item{power}{Scaling factor for the magnitude of differences between
communities (\eqn{p}). Default: \code{1.5}}

\item{pseudocount}{The value to add to all counts in \code{counts} to prevent
taking \code{log(0)} for unobserved features. The default, \code{NULL}, selects
the smallest non-zero value in \code{counts}.}

\item{alpha}{How much weight to give to relative abundances; a value
between 0 and 1, inclusive. Setting \code{alpha=1} is equivalent to
\code{normalized_unifrac()}.}

\item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
must be present in \code{tree}. Can be omitted if a tree is embedded with
the \code{counts} object or as \code{attr(counts, 'tree')}.}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\value{
A numeric vector.
}
\description{
Beta Diversity Wrapper Function
}
\details{
\strong{List of Beta Diversity Metrics}\tabular{ll}{
   Option / Function Name \tab Metric Name \cr
   \code{aitchison} \tab Aitchison distance \cr
   \code{bhattacharyya} \tab Bhattacharyya distance \cr
   \code{bray} \tab Bray-Curtis dissimilarity \cr
   \code{canberra} \tab Canberra distance \cr
   \code{chebyshev} \tab Chebyshev distance \cr
   \code{chord} \tab Chord distance \cr
   \code{clark} \tab Clark's divergence distance \cr
   \code{divergence} \tab Divergence \cr
   \code{euclidean} \tab Euclidean distance \cr
   \code{generalized_unifrac} \tab Generalized UniFrac (GUniFrac) \cr
   \code{gower} \tab Gower distance \cr
   \code{hamming} \tab Hamming distance \cr
   \code{hellinger} \tab Hellinger distance \cr
   \code{horn} \tab Horn-Morisita dissimilarity \cr
   \code{jaccard} \tab Jaccard distance \cr
   \code{jensen} \tab Jensen-Shannon distance \cr
   \code{jsd} \tab Jesen-Shannon divergence (JSD) \cr
   \code{lorentzian} \tab Lorentzian distance \cr
   \code{manhattan} \tab Manhattan distance \cr
   \code{matusita} \tab Matusita distance \cr
   \code{minkowski} \tab Minkowski distance \cr
   \code{morisita} \tab Morisita dissimilarity \cr
   \code{motyka} \tab Motyka dissimilarity \cr
   \code{normalized_unifrac} \tab Normalized Weighted UniFrac \cr
   \code{ochiai} \tab Otsuka-Ochiai dissimilarity \cr
   \code{psym_chisq} \tab Probabilistic Symmetric Chi-Squared distance \cr
   \code{soergel} \tab Soergel distance \cr
   \code{sorensen} \tab Dice-Sorensen dissimilarity \cr
   \code{squared_chisq} \tab Squared Chi-Squared distance \cr
   \code{squared_chord} \tab Squared Chord distance \cr
   \code{squared_euclidean} \tab Squared Euclidean distance \cr
   \code{topsoe} \tab Topsoe distance \cr
   \code{unweighted_unifrac} \tab Unweighted UniFrac \cr
   \code{variance_adjusted_unifrac} \tab Variance-Adjusted Weighted UniFrac (VAW-UniFrac) \cr
   \code{wave_hedges} \tab Wave Hedges distance \cr
   \code{weighted_unifrac} \tab Weighted UniFrac \cr
}


\strong{Flexible name matching}

Case insensitive and partial matching. Any runs of non-alpha characters are
converted to underscores. E.g. \verb{metric = 'Weighted UniFrac} selects
\code{weighted_unifrac}.

UniFrac names can be shortened to the first letter plus "unifrac". E.g.
\code{uunifrac}, \code{w_unifrac}, or \verb{V UniFrac}. These also support partial matching.

Finished code should always use the full primary option name to avoid
ambiguity with future additions to the metrics list.
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Bray-Curtis distances
    beta_div(ex_counts, 'bray')
    
    # Generalized UniFrac distances
    beta_div(ex_counts, 'GUniFrac', tree = ex_tree)
    
}
