% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{chao1}
\alias{chao1}
\title{Chao1 Richness Estimator}
\usage{
chao1(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A non-parametric estimator of the lower bound of species richness.
}
\details{
The Chao1 estimator uses the ratio of singletons to doubletons to estimate the number of missing species:
\deqn{n + \frac{(F_1)^2}{2 F_2}}

Where:
\itemize{
\item \eqn{n} : The number of observed features.
\item \eqn{F_1} : Number of features observed once (singletons).
\item \eqn{F_2} : Number of features observed twice (doubletons).
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
sum(x>0) + (sum(x == 1) ** 2) / (2 * sum(x == 2))
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    chao1(ex_counts)
}
\references{
Chao, A. (1984). Nonparametric estimation of the number of classes in a population. \emph{Scandinavian Journal of Statistics}, 11, 265-270.
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{observed}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
