% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_counts}
\alias{ex_counts}
\title{Example counts matrix}
\format{
A matrix of 4 samples (columns) x 6 genera (rows).
}
\source{
Derived from The Human Microbiome Project dataset.
\url{https://commonfund.nih.gov/hmp}
}
\usage{
ex_counts
}
\description{
Genera found on four human body sites.
}
\keyword{datasets}
