% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{faith}
\alias{faith}
\title{Faith's Phylogenetic Diversity (PD)}
\usage{
faith(counts, tree = NULL, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
must be present in \code{tree}. Can be omitted if a tree is embedded with
the \code{counts} object or as \code{attr(counts, 'tree')}.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Calculates the sum of the branch lengths for all species present in a sample.
}
\details{
Faith's PD is defined as:
\deqn{\sum_{i = 1}^{n} L_i A_i}

Where:
\itemize{
\item \eqn{n} : The number of branches in the phylogenetic tree.
\item \eqn{L_i} : The length of the \eqn{i}-th branch.
\item \eqn{A_i} : A binary value (1 if any descendants of branch \eqn{i} are present in the sample, 0 otherwise).
}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    faith(ex_counts, tree = ex_tree)
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. \emph{Biological Conservation}, 61(1), 1-10. \doi{10.1016/0006-3207(92)91201-3}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Phylogenetic metrics: 
\code{\link{generalized_unifrac}()},
\code{\link{normalized_unifrac}()},
\code{\link{unweighted_unifrac}()},
\code{\link{variance_adjusted_unifrac}()},
\code{\link{weighted_unifrac}()}
}
\concept{Phylogenetic metrics}
