% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{inv_simpson}
\alias{inv_simpson}
\title{Inverse Simpson}
\usage{
inv_simpson(counts, cpus = n_cpus())
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\value{
A numeric vector.
}
\description{
Inverse Simpson alpha diversity metric.
}
\details{
The Inverse Simpson index is another way to present the information from the
Simpson index, but in a more intuitive format. While the standard Simpson
index calculates the probability of two randomly selected individuals
belonging to the same species (where a lower value means more diversity), the
Inverse Simpson index is the reciprocal of that value (1/D). This simple
transformation makes the index easier to interpret: the value increases as
diversity increases.

The primary advantage of the Inverse Simpson index is that its value can be
understood as the "effective number of species". This means a community with
an Inverse Simpson index of 10 has a diversity that is equivalent to a
community composed of 10 equally abundant species. This provides a more
direct and biologically meaningful interpretation compared to more abstract
indices. Like the standard Simpson index, it is more heavily weighted by the
most abundant species and is less sensitive to sampling depth than richness
metrics like Observed Features or Chao1.
}
\section{Calculation}{


Pre-transformation: drop all OTUs with zero abundance.

In the formulas below, \eqn{x} is a single column (sample) from \code{counts}.
\eqn{p} are the relative abundances.

\deqn{p_{i} = \displaystyle \frac{x_i}{\sum x}}
\deqn{D = \displaystyle 1 / \sum_{i = 1}^{n} p_{i}\times\ln(p_{i})}

\if{html}{\out{<div class="sourceCode">}}\preformatted{  x <- c(4, 0, 3, 2, 6)[-2]  
  p <- x / sum(x)
  1 / sum(p * log(p))
  #>  -0.7636352
}\if{html}{\out{</div>}}
}

\examples{
    # Example counts matrix
    ex_counts
    
    # Inverse Simpson diversity values
    inv_simpson(ex_counts)
    
    # Low diversity
    inv_simpson(c(100, 1, 1, 1, 1)) # 1.08
    
    # High diversity
    inv_simpson(c(20, 20, 20, 20, 20)) # 5
    
    # Low richness
    inv_simpson(1:3) # 2.57
    
    # High richness
    inv_simpson(1:100) # 75.37
    
}
\references{
Simpson EH 1949.
Measurement of diversity.
Nature, 163.
\doi{10.1038/163688a0}
}
\seealso{
Other alpha_diversity: 
\code{\link{alpha_div}()},
\code{\link{beta_div}()},
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{observed}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
