% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{inv_simpson}
\alias{inv_simpson}
\title{Inverse Simpson alpha diversity metric}
\usage{
inv_simpson(counts, cpus = n_cpus())
}
\arguments{
\item{counts}{An OTU abundance matrix where each column is a sample, and
each row is an OTU. Any object coercible with \code{as.matrix()} can be
given here, as well as \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment},
and \code{TreeSummarizedExperiment} objects.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Inverse Simpson alpha diversity metric
}
\examples{
    # Example counts matrix
    ex_counts
    
    # Inverse Simpson diversity values
    inv_simpson(ex_counts)
    
}
\seealso{
Other alpha_diversity: 
\code{\link{chao1}()},
\code{\link{faith}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{alpha_diversity}
