% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{inv_simpson}
\alias{inv_simpson}
\title{Inverse Simpson Index}
\usage{
inv_simpson(counts, norm = "percent", margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{norm}{Normalize the incoming counts. Options are:
\describe{
\item{\code{norm = "percent"} - }{ Relative abundance (sample abundances sum to 1). }
\item{\code{norm = "binary"}  - }{ Unweighted presence/absence (each count is either 0 or 1). }
\item{\code{norm = "clr"}     - }{ Centered log ratio. }
\item{\code{norm = "none"}    - }{ No transformation. }
}
Default: \code{'percent'}, which is the expected input for these formulas.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A transformation of the Simpson index that represents the "effective number of species".
}
\details{
The Inverse Simpson index is defined as:
\deqn{1 / \sum_{i = 1}^{n} P_i^2}

Where:
\itemize{
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{p <- x / sum(x)
1 / sum(p ** 2)
}\if{html}{\out{</div>}}
}
\examples{
    inv_simpson(ex_counts)
}
\references{
Simpson, E. H. (1949). Measurement of diversity. \emph{Nature}, 163, 688. \doi{10.1038/163688a0}
}
\seealso{
alpha_div

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{fisher}()},
\code{\link{shannon}()},
\code{\link{simpson}()}
}
\concept{Diversity metrics}
