% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{jaccard}
\alias{jaccard}
\title{Jaccard distance}
\usage{
jaccard(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Measures dissimilarity between sample sets.
}
\details{
The Jaccard distance is defined as:
\deqn{1 - \frac{J}{(A + B - J)}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{1 - sum(x & y) / sum(x | y)
}\if{html}{\out{</div>}}
}
\examples{
    jaccard(ex_counts)
}
\references{
Jaccard, P. (1912). The distribution of the flora in the alpine zone. \emph{New Phytologist}, 11(2), 37-50. \doi{10.1111/j.1469-8137.1912.tb05611.x}

Jaccard, P. (1908). Nouvelles recherches sur la distribution florale. \emph{Bulletin de la Societe Vaudoise des Sciences Naturelles}, 44(163), 223-270. \doi{10.5169/seals-268384}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{ochiai}()},
\code{\link{sorensen}()}
}
\concept{Presence/Absence metrics}
