% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{jaccard}
\alias{jaccard}
\title{Jaccard distance}
\usage{
jaccard(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Measures dissimilarity between sample sets.
}
\details{
The Jaccard distance is defined as:
\deqn{1 - \frac{J}{(A + B - J)}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
y <- ex_counts[2,]
1 - sum(x & y) / sum(x | y)
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    jaccard(ex_counts)
}
\references{
Jaccard, P. (1912). The distribution of the flora in the alpine zone. \emph{New Phytologist}, 11(2), 37-50. \doi{10.1111/j.1469-8137.1912.tb05611.x}

Jaccard, P. (1908). Nouvelles recherches sur la distribution florale. \emph{Bulletin de la Societe Vaudoise des Sciences Naturelles}, 44(163), 223-270. \doi{10.5169/seals-268384}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{ochiai}()},
\code{\link{sorensen}()}
}
\concept{Presence/Absence metrics}
