% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{margalef}
\alias{margalef}
\title{Margalef's Richness Index}
\usage{
margalef(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A richness metric that normalizes the number of species by the log of the total sample size.
}
\details{
Margalef's index is defined as:
\deqn{\frac{n - 1}{\ln{X_T}}}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{X_T} : Total of all counts (sequencing depth).
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{(sum(x > 0) - 1) / log(sum(x))
}\if{html}{\out{</div>}}
}
\examples{
    margalef(ex_counts)
}
\references{
Margalef, R. (1958). Information theory in ecology. \emph{General Systems}, 3, 36-71.
}
\seealso{
alpha_div

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{chao1}()},
\code{\link{menhinick}()},
\code{\link{observed}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
