% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{mcintosh}
\alias{mcintosh}
\title{McIntosh Index}
\usage{
mcintosh(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A dominance index based on the Euclidean distance from the origin.
}
\details{
The McIntosh index is defined as:
\deqn{\frac{X_T - \sqrt{\sum_{i = 1}^{n} (X_i)^2}}{X_T - \sqrt{X_T}}}

Where:
\itemize{
\item \eqn{X_i} : Integer count of the \eqn{i}-th feature.
\item \eqn{X_T} : Total of all counts.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{(sum(x) - sqrt(sum(x^2))) / (sum(x) - sqrt(sum(x)))
}\if{html}{\out{</div>}}
}
\examples{
    mcintosh(ex_counts)
}
\references{
McIntosh, R. P. (1967). An index of diversity and the relation of certain concepts to diversity. \emph{Ecology}, 48(3), 392-404. \doi{10.2307/1932674}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Dominance metrics: 
\code{\link{berger}()}
}
\concept{Dominance metrics}
