% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{metrics}
\alias{metrics}
\title{API for browsing functions}
\format{
An object of class \code{list} of length 2.
}
\usage{
metrics
}
\value{
A nested list of functions for alpha and beta diversity.
}
\description{
Programmatic access to the lists of available metrics, and their associated
functions.
}
\examples{
    names(metrics)
    
    names(metrics$alpha)
    
    names(formals(metrics$alpha$faith))
    
    metrics$alpha$faith(ex_counts, ex_tree)

}
\keyword{datasets}
