% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{n_cpus}
\alias{n_cpus}
\title{Number of CPU Cores}
\usage{
n_cpus()
}
\value{
A scalar integer, guaranteed to be at least \code{1}.
}
\description{
A thin wrapper around
\code{parallel::detectCores(all.tests = TRUE, logical = TRUE)} which falls back
to \code{1} when the number of CPU cores cannot be detected, or when the system
does not support \code{pthreads}. Consider using \code{parallely::availableCores()}
in place of \code{n_cpus()} for more advanced interrogation of system resources.
}
\examples{
    n_cpus()

}
