% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{observed}
\alias{observed}
\title{Observed Features}
\usage{
observed(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The count of unique features (richness) in a sample.
}
\details{
Observed features is defined simply as the number of features with non-zero abundance:
\deqn{n}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum(x > 0)
}\if{html}{\out{</div>}}
}
\examples{
    observed(ex_counts)
}
\seealso{
alpha_div

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{chao1}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
