% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{observed}
\alias{observed}
\title{Observed Features}
\usage{
observed(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The count of unique features (richness) in a sample.
}
\details{
Observed features is defined simply as the number of features with non-zero abundance:
\deqn{n}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
sum(x > 0)
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    observed(ex_counts)
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{chao1}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{squares}()}
}
\concept{Richness metrics}
