% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{ochiai}
\alias{ochiai}
\title{Otsuka-Ochiai dissimilarity}
\usage{
ochiai(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
Also known as the cosine similarity for binary data.
}
\details{
The Otsuka-Ochiai dissimilarity is defined as:
\deqn{1 - \frac{J}{\sqrt{AB}}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{1 - sum(x & y) / sqrt(sum(x>0) * sum(y>0)) 
}\if{html}{\out{</div>}}
}
\examples{
    ochiai(ex_counts)
}
\references{
Ochiai, A. (1957). Zoogeographic studies on the soleoid fishes found in Japan and its neighbouring regions. \emph{Bulletin of the Japanese Society of Scientific Fisheries}, 22, 526-530.
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{jaccard}()},
\code{\link{sorensen}()}
}
\concept{Presence/Absence metrics}
