% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{psym_chisq}
\alias{psym_chisq}
\title{Probabilistic Symmetric Chi-Squared distance}
\usage{
psym_chisq(
  counts,
  norm = "percent",
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{norm}{Normalize the incoming counts. Options are:
\itemize{
\item \code{'percent'}: Relative abundance (sample abundances sum to 1).
\item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
\item \code{'clr'}: Centered log ratio.
\item \code{'none'}: No transformation.
}

Default: \code{'percent'}, which is the expected input for these formulas.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A chi-squared based distance metric for comparing probability distributions.
}
\details{
The Probabilistic Symmetric \eqn{\chi^2} distance is defined as:
\deqn{2\sum_{i=1}^{n}\frac{(P_i - Q_i)^2}{P_i + Q_i}}

Where:
\itemize{
\item \eqn{P_i}, \eqn{Q_i} : Proportional abundances of the \eqn{i}-th feature.
\item \eqn{n} : The number of features.
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{2 * sum((x - y)^2 / (x + y))
}\if{html}{\out{</div>}}
}
\examples{
    psym_chisq(ex_counts)
}
\references{
Cha, S.-H. (2007). Comprehensive survey on distance/similarity measures between probability density functions. \emph{International Journal of Mathematical Models and Methods in Applied Sciences}, 1(4), 300–307.
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Abundance metrics: 
\code{\link{aitchison}()},
\code{\link{bhattacharyya}()},
\code{\link{bray}()},
\code{\link{canberra}()},
\code{\link{chebyshev}()},
\code{\link{chord}()},
\code{\link{clark}()},
\code{\link{divergence}()},
\code{\link{euclidean}()},
\code{\link{gower}()},
\code{\link{hellinger}()},
\code{\link{horn}()},
\code{\link{jensen}()},
\code{\link{jsd}()},
\code{\link{lorentzian}()},
\code{\link{manhattan}()},
\code{\link{matusita}()},
\code{\link{minkowski}()},
\code{\link{morisita}()},
\code{\link{motyka}()},
\code{\link{soergel}()},
\code{\link{squared_chisq}()},
\code{\link{squared_chord}()},
\code{\link{squared_euclidean}()},
\code{\link{topsoe}()},
\code{\link{wave_hedges}()}
}
\concept{Abundance metrics}
