% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{shannon}
\alias{shannon}
\title{Shannon Diversity Index}
\usage{
shannon(counts, norm = "percent", margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{norm}{Normalize the incoming counts. Options are:
\itemize{
\item \code{'percent'}: Relative abundance (sample abundances sum to 1).
\item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
\item \code{'clr'}: Centered log ratio.
\item \code{'none'}: No transformation.
}

Default: \code{'percent'}, which is the expected input for these formulas.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A commonly used diversity index accounting for both abundance and evenness.
}
\details{
The Shannon index (entropy) is defined as:
\deqn{-\sum_{i = 1}^{n} P_i \times \ln(P_i)}

Where:
\itemize{
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{p <- x / sum(x)
-sum(p * log(p))
}\if{html}{\out{</div>}}
}
\examples{
    shannon(ex_counts)
}
\references{
Shannon, C. E. (1948). A mathematical theory of communication. \emph{Bell System Technical Journal}, 27, 379-423.

Shannon, C. E., & Weaver, W. (1949). \emph{The Mathematical Theory of Communication}. University of Illinois Press.
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{fisher}()},
\code{\link{inv_simpson}()},
\code{\link{simpson}()}
}
\concept{Diversity metrics}
