% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{shannon}
\alias{shannon}
\title{Shannon Diversity Index}
\usage{
shannon(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A commonly used diversity index accounting for both abundance and evenness.
}
\details{
The Shannon index (entropy) is defined as:
\deqn{-\sum_{i = 1}^{n} P_i \times \ln(P_i)}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
p <- x / sum(x)
-sum(p * log(p))
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    shannon(ex_counts)
}
\references{
Shannon, C. E. (1948). A mathematical theory of communication. \emph{Bell System Technical Journal}, 27, 379-423.

Shannon, C. E., & Weaver, W. (1949). \emph{The Mathematical Theory of Communication}. University of Illinois Press.
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{fisher}()},
\code{\link{inv_simpson}()},
\code{\link{simpson}()}
}
\concept{Diversity metrics}
