% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{simpson}
\alias{simpson}
\title{Gini-Simpson Index}
\usage{
simpson(counts, norm = "percent", margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{norm}{Normalize the incoming counts. Options are:
* \code{'percent'}: Relative abundance (sample abundances sum to 1).
* \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
* \code{'clr'}: Centered log ratio.
* \code{'none'}: No transformation.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   Default: `'percent'`, which is the expected input for these formulas.
}\if{html}{\out{</div>}}}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The probability that two entities taken at random from the dataset represent different types.
}
\details{
The Gini-Simpson index is defined as:
\deqn{1 - \sum_{i = 1}^{n} P_i^2}

Where:
\itemize{
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{p <- x / sum(x)
1 - sum(p ** 2)
}\if{html}{\out{</div>}}
}
\examples{
    simpson(ex_counts)
}
\references{
Simpson, E. H. (1949). Measurement of diversity. \emph{Nature}, 163, 688. \doi{10.1038/163688a0}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{fisher}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()}
}
\concept{Diversity metrics}
