% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{simpson}
\alias{simpson}
\title{Gini-Simpson Index}
\usage{
simpson(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
The probability that two entities taken at random from the dataset represent different types.
}
\details{
The Gini-Simpson index is defined as:
\deqn{1 - \sum_{i = 1}^{n} P_i^2}

Where:
\itemize{
\item \eqn{n} : The number of features.
\item \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
p <- x / sum(x)
1 - sum(p ** 2)
}\if{html}{\out{</div>}}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    simpson(ex_counts)
}
\references{
Simpson, E. H. (1949). Measurement of diversity. \emph{Nature}, 163, 688. \doi{10.1038/163688a0}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Diversity metrics: 
\code{\link{brillouin}()},
\code{\link{fisher}()},
\code{\link{inv_simpson}()},
\code{\link{shannon}()}
}
\concept{Diversity metrics}
