% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{sorensen}
\alias{sorensen}
\title{Dice-Sorensen dissimilarity}
\usage{
sorensen(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A statistic used for comparing the similarity of two samples.
}
\details{
The Dice-Sorensen dissimilarity is defined as:
\deqn{\frac{2J}{(A + B)}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{2 * sum(x & y) / sum(x>0, y>0)
}\if{html}{\out{</div>}}
}
\examples{
    sorensen(ex_counts)
}
\references{
Sørensen, T. (1948). A method of establishing groups of equal amplitude in plant sociology based on similarity of species content. \emph{Kongelige Danske Videnskabernes Selskab, Biologiske Skrifter}, 5, 1-34.

Dice, L. R. (1945). Measures of the amount of ecologic association between species. \emph{Ecology}, 26(3), 297–302. \doi{10.2307/1932409}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{jaccard}()},
\code{\link{ochiai}()}
}
\concept{Presence/Absence metrics}
