% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{sorensen}
\alias{sorensen}
\title{Dice-Sorensen dissimilarity}
\usage{
sorensen(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A statistic used for comparing the similarity of two samples.
}
\details{
The Dice-Sorensen dissimilarity is defined as:
\deqn{\frac{2J}{(A + B)}}

Where:
\itemize{
\item \eqn{A}, \eqn{B} : Number of features in each sample.
\item \eqn{J} : Number of features in common (intersection).
}

Base R Equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{2 * sum(x & y) / sum(x>0, y>0)
}\if{html}{\out{</div>}}
}
\examples{
    sorensen(ex_counts)
}
\references{
Sørensen, T. (1948). A method of establishing groups of equal amplitude in plant sociology based on similarity of species content. \emph{Kongelige Danske Videnskabernes Selskab, Biologiske Skrifter}, 5, 1-34.
}
\seealso{
beta_div

Other Presence/Absence metrics: 
\code{\link{hamming}()},
\code{\link{jaccard}()},
\code{\link{ochiai}()}
}
\concept{Presence/Absence metrics}
