% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.r
\name{squares}
\alias{squares}
\title{Squares Richness Estimator}
\usage{
squares(counts, margin = 1L, cpus = n_cpus())
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
optimizing large datasets.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
special object class (e.g. \code{phyloseq}). Default: \code{1L}}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A richness estimator based on the concept of "squares" (counts of species observed once or twice).
}
\details{
The Squares estimator is defined as:
\deqn{n + \frac{(F_1)^2 \sum_{i=1}^{n} (X_i)^2}{X_T^2 - nF_1}}

Where:
\itemize{
\item \eqn{n} : The number of observed features.
\item \eqn{X_T} : Total of all counts.
\item \eqn{F_1} : Number of features observed once (singletons).
\item \eqn{X_i} : Integer count of the \eqn{i}-th feature.
}

\strong{Base R Equivalent:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{N  <- sum(x)      # sampling depth
S  <- sum(x > 0)  # observed features
F1 <- sum(x == 1) # singletons
S + ((sum(x^2) * (F1^2)) / ((N^2) - F1 * S))
}\if{html}{\out{</div>}}
}
\examples{
    squares(ex_counts)
}
\references{
Alroy, J. (2018). Limits to species richness estimates based on subsampling. \emph{Paleobiology}, 44(2), 177-194. \doi{10.1017/pab.2017.38}
}
\seealso{
\code{alpha_div()}, \code{vignette('adiv')}

Other Richness metrics: 
\code{\link{ace}()},
\code{\link{chao1}()},
\code{\link{margalef}()},
\code{\link{menhinick}()},
\code{\link{observed}()}
}
\concept{Richness metrics}
