% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{unweighted_unifrac}
\alias{unweighted_unifrac}
\title{Unweighted UniFrac}
\usage{
unweighted_unifrac(
  counts,
  tree = NULL,
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
Typically contains absolute abundances (integer counts), though
proportions are also accepted.}

\item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
must be present in \code{tree}. Can be omitted if a tree is embedded with
the \code{counts} object or as \code{attr(counts, 'tree')}.}

\item{margin}{The margin containing samples. \code{1} if samples are rows,
\code{2} if samples are columns. Ignored when \code{counts} is a special object
class (e.g. \code{phyloseq}). Default: \code{1}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A phylogenetic distance metric that accounts for the presence/absence of lineages.
}
\details{
The Unweighted UniFrac distance is defined as:
\deqn{\frac{1}{n}\sum_{i=1}^{n} L_i|A_i - B_i|}

Where:
\itemize{
\item \eqn{n} : The number of branches in the tree.
\item \eqn{L_i} : The length of the \eqn{i}-th branch.
\item \eqn{A_i}, \eqn{B_i} : Binary values (0 or 1) indicating if descendants of branch \eqn{i} are present in sample A or B.
}
}
\section{Input Types}{


The \code{counts} parameter is designed to accept a simple numeric matrix, but
seamlessly supports objects from the following biological data packages:
\itemize{
\item \code{phyloseq}
\item \code{rbiom}
\item \code{SummarizedExperiment}
\item \code{TreeSummarizedExperiment}
}

For large datasets, standard matrix operations may be slow. See
\code{vignette('performance')} for details on using optimized formats
(e.g. sparse matrices) and parallel processing.
}

\examples{
    unweighted_unifrac(ex_counts, tree = ex_tree)
}
\references{
Lozupone, C., & Knight, R. (2005). UniFrac: a new phylogenetic method for comparing microbial communities. \emph{Applied and Environmental Microbiology}, 71(12), 8228-8235. \doi{10.1128/AEM.71.12.8228-8235.2005}
}
\seealso{
\code{beta_div()}, \code{vignette('bdiv')}, \code{vignette('bdiv_guide')}

Other Phylogenetic metrics: 
\code{\link{faith}()},
\code{\link{generalized_unifrac}()},
\code{\link{normalized_unifrac}()},
\code{\link{variance_adjusted_unifrac}()},
\code{\link{weighted_unifrac}()}
}
\concept{Phylogenetic metrics}
