% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.r
\name{variance_adjusted_unifrac}
\alias{variance_adjusted_unifrac}
\title{Variance-Adjusted Weighted UniFrac}
\usage{
variance_adjusted_unifrac(
  counts,
  tree = NULL,
  margin = 1L,
  pairs = NULL,
  cpus = n_cpus()
)
}
\arguments{
\item{counts}{A numeric matrix of count data where each column is a
feature, and each row is a sample. Any object coercible with
\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
optimal performance with very large datasets, see the guide in
\code{vignette('performance')}.}

\item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
must be present in \code{tree}. Can be omitted if a tree is embedded with
the \code{counts} object or as \code{attr(counts, 'tree')}.}

\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
\code{TreeSummarizedExperiment} object. Default: \code{1L}}

\item{pairs}{Which combinations of samples should distances be
calculated for? The default value (\code{NULL}) calculates all-vs-all.
Provide a numeric or logical vector specifying positions in the
distance matrix to calculate. See examples.}

\item{cpus}{How many parallel processing threads should be used. The
default, \code{n_cpus()}, will use all logical CPU cores.}
}
\description{
A weighted UniFrac that adjusts for the expected variance of the metric.
}
\details{
The Variance-Adjusted Weighted UniFrac distance is defined as:
\deqn{\frac{\sum_{i=1}^{n} L_i\frac{|P_i - Q_i|}{\sqrt{(P_i + Q_i)(2 - P_i - Q_i)}} }{\sum_{i=1}^{n} L_i\frac{P_i + Q_i}{\sqrt{(P_i + Q_i)(2 - P_i - Q_i)}} }}

Where:
\itemize{
\item \eqn{n} : The number of branches in the tree.
\item \eqn{L_i} : The length of the \eqn{i}-th branch.
\item \eqn{P_i}, \eqn{Q_i} : The proportion of the community descending from branch \eqn{i} in sample P and Q.
}
}
\examples{
    variance_adjusted_unifrac(ex_counts, tree = ex_tree)
}
\references{
Chang, Q., Luan, Y., & Sun, F. (2011). Variance adjusted weighted UniFrac: a powerful beta diversity measure for comparing communities based on phylogeny. \emph{BMC Bioinformatics}, 12, 118. \doi{10.1186/1471-2105-12-118}
}
\seealso{
beta_div

Other Phylogenetic metrics: 
\code{\link{faith}()},
\code{\link{generalized_unifrac}()},
\code{\link{normalized_unifrac}()},
\code{\link{unweighted_unifrac}()},
\code{\link{weighted_unifrac}()}
}
\concept{Phylogenetic metrics}
